Fix copypaste mistakes and add "post" version tests.
