+++ src/NClient.Providers/NClient.Providers.Serialization.Protobuf.ProtobufNet/ProtobufSerializerProvider.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Logging;
 using NClient.Common.Helpers;
 
-namespace NClient.Providers.Serialization.Protobuf
+namespace NClient.Providers.Serialization.Protobuf.ProtobufNet
 {
     public class ProtobufSerializerProvider : ISerializerProvider
     {
