+++ tests/NClient.Packages/NClient.Packages.Tests/NClient.Packages.Tests.csproj
@@ -28,6 +28,7 @@
         <PackageReference Include="NClient.Providers.Serialization.Json.System" Version="*-*"/>
         <PackageReference Include="NClient.Providers.Serialization.Json.Newtonsoft" Version="*-*"/>
         <PackageReference Include="NClient.Providers.Serialization.Xml.System" Version="*-*"/>
+        <PackageReference Include="NClient.Providers.Serialization.Protobuf.ProtobufNet" Version="*-*"/>
         <PackageReference Include="NClient.Extensions.DependencyInjection" Version="*-*"/>
     </ItemGroup>
