+++ tests/NClient.Providers/NClient.Providers.Serialization.Protobuf.ProtobufNet.Tests/Models/Point.cs
@@ -1,6 +1,6 @@
 using ProtoBuf;
 
-namespace NClient.Providers.Serialization.Protobuf.Tests.Models
+namespace NClient.Providers.Serialization.Protobuf.ProtobufNet.Tests.Models
 {
     [ProtoContract]
     public struct Point
